feat: install the spanish version of the validations
first we need to run de command: "php artisan lang:publish" if we don't have the lang directory on laravel.
that command will add the directory with the english version, but for spanish we need to go to github and download the repository: https://github.com/Laraveles/spanish and extract the "es" directory on the resources/lang/es folder and copy and paste on our project, on the same folder (/lang).

finaly we need to update the line  "'locale' => 'en'" to  "'locale' => 'es'" from the config/app.php file.
this way we're telling laravel to use the spanish folder we already added